+++ browserstep/debug.py
@@ -1,0 +1,6 @@
+from behave import *
+
+
+@step('I debug')
+def step_impl(context):
+    import pdb; pdb.set_trace()